2021-03-25
Fixed the user and creator error. Only the creator can modify the posts.

Todo:
Fix the like system.
When you press the like buttom, link it to create a like with the user_id and the recipe_id.
Count the amount of elements of likes that the recipe_id is in.